fix: 잉크 토스트가 안 뜨는 오류 수정
